Load prices from CSV
