+++ custom_components/som_energia/price/prices.csv
@@ -1,0 +1,7 @@
+Inicio Periodo,Final Periodo,Compensación,Punta,Llano,Valle
+2022-01-01,2022-01-31,0.000,0.396,0.286,0.228
+2023-01-01,2023-04-30,0.000,0.342,0.281,0.234
+2023-05-01,2023-12-31,0.130,0.295,0.237,0.199
+2024-01-01,2024-03-31,0.070,0.247,0.189,0.154
+2024-04-01,2024-10-31,0.060,0.212,0.154,0.118
+2024-11-01,2999-12-31,0.050,0.226,0.166,0.127
